backend/requirements.txt
fastapi
uvicorn
python-multipart
pymupdf
easyocr
edge-tts
pydantic
numpy
Pillow
